Refactored status and utils to separate files. Correct response for method not allowed and if application/json not in accept header of request
