+++ constants.py
@@ -5,3 +5,5 @@ google_auth_endpoint = "https://accounts.google.com/o/oauth2/v2/auth"
 testjwt = "eyJhbGciOiJSUzI1NiIsImtpZCI6IjQ5MjcxMGE3ZmNkYjE1Mzk2MGNlMDFmNzYwNTIwYTMyYzg0NTVkZmYiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJhY2NvdW50cy5nb29nbGUuY29tIiwiYXpwIjoiODcwMTU1MzgzMzQ0LWwzc2NxZW5zbmJmOHRpb3F2Mm82MG9yZmUxZXI0NGZ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiYXVkIjoiODcwMTU1MzgzMzQ0LWwzc2NxZW5zbmJmOHRpb3F2Mm82MG9yZmUxZXI0NGZ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwic3ViIjoiMTE3Mzk1NTU1Nzk1NDU2ODM0MDIwIiwiaGQiOiJvcmVnb25zdGF0ZS5lZHUiLCJlbWFpbCI6InN0YW5sZW1pQG9yZWdvbnN0YXRlLmVkdSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJhdF9oYXNoIjoiaFN5WTRwSkM0ZGZ5YUR4WkZLLUVqUSIsImlhdCI6MTU5MTU2MDc0OCwiZXhwIjoxNTkxNTY0MzQ4fQ.oAW9EdRb4Hsz5cQ2s75FN84ShAaLv2dPc9gmeV6MYvZzecC2Xgmk_jc1Lr0hkLz0o0u3haoTUaPUy3DZ4EeqdM0njpducbmrOW6Y8MAVMy99fpSu2Aw0VN3mGcEl-zqZgBsWUOgdG_bP5QOTAcLQ-gqFKD2rjDOO_6iOvedXyxfW8nI7Nhj1zcALl5u8HTjLL-8NHOzoQ2jY_fih8DT-oP1pybYD9hgrQx5lkTa7Wj6Hu7pEFnOY1bFXCkL5-wOGfGivgpkyrMfarO-2UF1IgRm2VE0OIjjvhFEqPEJp7EAXZMQxQCxQLRhsfPm1fJBMh1Up8L56E_q4gx30tWW_HA"
 testjwt2 = "eyJhbGciOiJSUzI1NiIsImtpZCI6IjQ5MjcxMGE3ZmNkYjE1Mzk2MGNlMDFmNzYwNTIwYTMyYzg0NTVkZmYiLCJ0eXAiOiJKV1QifQ.eyJpc3MiOiJhY2NvdW50cy5nb29nbGUuY29tIiwiYXpwIjoiODcwMTU1MzgzMzQ0LWwzc2NxZW5zbmJmOHRpb3F2Mm82MG9yZmUxZXI0NGZ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwiYXVkIjoiODcwMTU1MzgzMzQ0LWwzc2NxZW5zbmJmOHRpb3F2Mm82MG9yZmUxZXI0NGZ0LmFwcHMuZ29vZ2xldXNlcmNvbnRlbnQuY29tIiwic3ViIjoiMTE3Mzk1NTU1Nzk1NDU2ODM0MDIwIiwiaGQiOiJvcmVnb25zdGF0ZS5lZHUiLCJlbWFpbCI6InN0YW5sZW1pQG9yZWdvbnN0YXRlLmVkdSIsImVtYWlsX3ZlcmlmaWVkIjp0cnVlLCJhdF9oYXNoIjoiOXoyNUgtemkzb3RndjhtS0YycGhWQSIsImlhdCI6MTU5MTU2MDk2NCwiZXhwIjoxNTkxNTY0NTY0fQ.plZSrWXwQ8XBi0IChJqdkqlPKlLnNlzAJR7LRUrtX4nc9lsXuiu-2F1DH8tb2dCBwA2bX9sTX3k5eWCgRZluvLscsXNM77KgR4mMJUvtjn05FDOc0QZptLtxeXFDSbxLN8kxiT8eh9YQtkTebjAecQVqK1_DG8zfPBb81QRToCitVfQA6Krv2Q3j6Tb3iTsqhHIyQWZhlaumgayBWJacOAPU7ok_fiBOJq_mPBtK5cb9jX31-kQtZN28OhhiBde3-Dv3IzAjHx_rVuEvTq_rAsOy_NBH3Miu3X8S9SFiQWQlT2kyXrPuQHNZ8bsAASOjjlgs8aIqaL5fQ9tqsm4uyw"
 users = "users"
+json_mimetype = "application/json"
+all_mimetype = "*/*"
